+++ src/lib.ts
@@ -5,6 +5,7 @@ import {
   cliExecute,
   eat,
   Familiar,
+  familiarWeight,
   fileToBuffer,
   fullnessLimit,
   gametimeToInt,
@@ -13,6 +14,7 @@ import {
   gitAtHead,
   gitInfo,
   handlingChoice,
+  haveEquipped,
   haveSkill,
   inebrietyLimit,
   isDarkMode,
@@ -23,6 +25,8 @@ import {
   meatDropModifier,
   Monster,
   mpCost,
+  myBjornedFamiliar,
+  myEnthronedFamiliar,
   myFamiliar,
   myFullness,
   myHp,
@@ -32,6 +36,7 @@ import {
   myMaxmp,
   myMp,
   mySoulsauce,
+  myThrall,
   myTurncount,
   numericModifier,
   print,
@@ -62,6 +67,7 @@ import {
   $monster,
   $skill,
   $slot,
+  $thralls,
   ActionSource,
   bestLibramToCast,
   ChateauMantegna,
@@ -82,6 +88,7 @@ import {
   realmAvailable,
   set,
   SongBoom,
+  SourceTerminal,
   sum,
   uneffect,
 } from "libram";
@@ -666,4 +673,112 @@ export function printEventLog(): void {
       HIGHLIGHT,
     );
   }
+}
+
+function untangleDigitizes(turnCount: number, chunks: number): number {
+  const turnsPerChunk = turnCount / chunks;
+  const monstersPerChunk = Math.sqrt((turnsPerChunk + 3) / 5 + 1 / 4) - 1 / 2;
+  return Math.round(chunks * monstersPerChunk);
+}
+
+export function digitizedMonstersRemainingForTurns(estimatedTurns: number): number {
+  if (!SourceTerminal.have()) return 0;
+
+  const digitizesLeft = SourceTerminal.getDigitizeUsesRemaining();
+  if (digitizesLeft === SourceTerminal.getMaximumDigitizeUses()) {
+    return untangleDigitizes(estimatedTurns, SourceTerminal.getMaximumDigitizeUses());
+  }
+
+  const monsterCount = SourceTerminal.getDigitizeMonsterCount() + 1;
+
+  const turnsLeftAtNextMonster = estimatedTurns - Counter.get("Digitize Monster");
+  if (turnsLeftAtNextMonster <= 0) return 0;
+  const turnsAtLastDigitize = turnsLeftAtNextMonster + ((monsterCount + 1) * monsterCount * 5 - 3);
+  return (
+    untangleDigitizes(turnsAtLastDigitize, digitizesLeft + 1) -
+    SourceTerminal.getDigitizeMonsterCount()
+  );
+}
+
+function maxCarriedFamiliarDamage(familiar: Familiar): number {
+  // Only considering familiars we reasonably may carry
+  switch (familiar) {
+    // +5 to Familiar Weight
+    case $familiar`Animated Macaroni Duck`:
+      return 50;
+    case $familiar`Barrrnacle`:
+    case $familiar`Gelatinous Cubeling`:
+    case $familiar`Penguin Goodfella`:
+      return 30;
+    case $familiar`Misshapen Animal Skeleton`:
+      return 40 + numericModifier("Spooky Damage");
+
+    // +25% Meat from Monsters
+    case $familiar`Hobo Monkey`:
+      return 25;
+
+    // +20% Meat from Monsters
+    case $familiar`Grouper Groupie`:
+      // Double sleaze damage at Barf Mountain
+      return (
+        25 + numericModifier("Sleaze Damage") * (myLocation() === $location`Barf Mountain` ? 2 : 1)
+      );
+    case $familiar`Jitterbug`:
+      return 20;
+    case $familiar`Mutant Cactus Bud`:
+      // 25 poison damage (25+12+6+3+1)
+      return 47;
+    case $familiar`Robortender`:
+      return 20;
+  }
+
+  return 0;
+}
+
+function maxFamiliarDamage(familiar: Familiar): number {
+  switch (familiar) {
+    case $familiar`Cocoabo`:
+      return familiarWeight(familiar) + 3;
+    case $familiar`Feather Boa Constrictor`:
+      // Double sleaze damage at Barf Mountain
+      return (
+        familiarWeight(familiar) +
+        3 +
+        numericModifier("Sleaze Damage") * (myLocation() === $location`Barf Mountain` ? 2 : 1)
+      );
+    case $familiar`Ninja Pirate Zombie Robot`:
+      return Math.floor((familiarWeight(familiar) + 3) * 1.5);
+  }
+  return 0;
+}
+
+export function maxPassiveDamage(): number {
+  // Only considering passive damage sources we reasonably may have
+  const vykeaMaxDamage =
+    get("_VYKEACompanionLevel") > 0 ? 10 * get("_VYKEACompanionLevel") + 10 : 0;
+
+  // Lasagmbie does max 2*level damage while Vermincelli does max level + (1/2 * level) + (1/2 * 1/2 * level) + ...
+  const thrallMaxDamage =
+    myThrall().level >= 5 && $thralls`Lasagmbie,Vermincelli`.includes(myThrall())
+      ? myThrall().level * 2
+      : 0;
+
+  const crownMaxDamage = haveEquipped($item`Crown of Thrones`)
+    ? maxCarriedFamiliarDamage(myEnthronedFamiliar())
+    : 0;
+
+  const bjornMaxDamage = haveEquipped($item`Buddy Bjorn`)
+    ? maxCarriedFamiliarDamage(myBjornedFamiliar())
+    : 0;
+
+  const familiarMaxDamage = maxFamiliarDamage(myFamiliar());
+
+  return vykeaMaxDamage + thrallMaxDamage + crownMaxDamage + bjornMaxDamage + familiarMaxDamage;
+}
+
+let monsterManuelCached: boolean | undefined = undefined;
+export function monsterManuelAvailable(): boolean {
+  if (monsterManuelCached !== undefined) return Boolean(monsterManuelCached);
+  monsterManuelCached = visitUrl("questlog.php?which=3").includes("Monster Manuel");
+  return Boolean(monsterManuelCached);
 }
